+++ src/main/java/facade/ICompanyFacade.java
@@ -11,7 +11,7 @@ public interface ICompanyFacade
 {
     public Company addCompany(Company c);
     public Company editCompany(Company c);
-    public Company deleteCompany(int id);
+    public Company deleteCompany(long id);
     public Company getCompany(int cvr);
     public Company getCompany(String number);
     public Company getCompany(String table, String data);
